# Undo refactoring in XOR encoding