--- a/tsdb/chunkenc/histogram.go
+++ b/tsdb/chunkenc/histogram.go
@@ -16,6 +16,7 @@ package chunkenc
 import (
 	"encoding/binary"
 	"math"
+	"math/bits"
 
 	"github.com/prometheus/prometheus/model/histogram"
 	"github.com/prometheus/prometheus/model/value"
@@ -874,4 +875,112 @@ func (it *histogramIterator) readSum() bool {
 	}
 	it.sum, it.leading, it.trailing = sum, leading, trailing
 	return true
+}
+
+func xorWrite(
+	b *bstream,
+	newValue, currentValue float64,
+	currentLeading, currentTrailing uint8,
+) (newLeading, newTrailing uint8) {
+	delta := math.Float64bits(newValue) ^ math.Float64bits(currentValue)
+
+	if delta == 0 {
+		b.writeBit(zero)
+		return currentLeading, currentTrailing
+	}
+	b.writeBit(one)
+
+	newLeading = uint8(bits.LeadingZeros64(delta))
+	newTrailing = uint8(bits.TrailingZeros64(delta))
+
+	// Clamp number of leading zeros to avoid overflow when encoding.
+	if newLeading >= 32 {
+		newLeading = 31
+	}
+
+	if currentLeading != 0xff && newLeading >= currentLeading && newTrailing >= currentTrailing {
+		// In this case, we stick with the current leading/trailing.
+		b.writeBit(zero)
+		b.writeBits(delta>>currentTrailing, 64-int(currentLeading)-int(currentTrailing))
+		return currentLeading, currentTrailing
+	}
+
+	b.writeBit(one)
+	b.writeBits(uint64(newLeading), 5)
+
+	// Note that if newLeading == newTrailing == 0, then sigbits == 64. But
+	// that value doesn't actually fit into the 6 bits we have.  Luckily, we
+	// never need to encode 0 significant bits, since that would put us in
+	// the other case (vdelta == 0).  So instead we write out a 0 and adjust
+	// it back to 64 on unpacking.
+	sigbits := 64 - newLeading - newTrailing
+	b.writeBits(uint64(sigbits), 6)
+	b.writeBits(delta>>newTrailing, int(sigbits))
+	return
+}
+
+func xorRead(
+	br *bstreamReader, currentValue float64, currentLeading, currentTrailing uint8,
+) (newValue float64, newLeading, newTrailing uint8, err error) {
+	var bit bit
+	var bits uint64
+
+	bit, err = br.readBitFast()
+	if err != nil {
+		bit, err = br.readBit()
+	}
+	if err != nil {
+		return
+	}
+	if bit == zero {
+		return currentValue, currentLeading, currentTrailing, nil
+	}
+	bit, err = br.readBitFast()
+	if err != nil {
+		bit, err = br.readBit()
+	}
+	if err != nil {
+		return
+	}
+	if bit == zero {
+		// Reuse leading/trailing zero bits.
+		newLeading, newTrailing = currentLeading, currentTrailing
+	} else {
+		bits, err = br.readBitsFast(5)
+		if err != nil {
+			bits, err = br.readBits(5)
+		}
+		if err != nil {
+			return
+		}
+		newLeading = uint8(bits)
+
+		bits, err = br.readBitsFast(6)
+		if err != nil {
+			bits, err = br.readBits(6)
+		}
+		if err != nil {
+			return
+		}
+		mbits := uint8(bits)
+		// 0 significant bits here means we overflowed and we actually
+		// need 64; see comment in xrWrite.
+		if mbits == 0 {
+			mbits = 64
+		}
+		newTrailing = 64 - newLeading - mbits
+	}
+
+	mbits := 64 - newLeading - newTrailing
+	bits, err = br.readBitsFast(mbits)
+	if err != nil {
+		bits, err = br.readBits(mbits)
+	}
+	if err != nil {
+		return
+	}
+	vbits := math.Float64bits(currentValue)
+	vbits ^= bits << newTrailing
+	newValue = math.Float64frombits(vbits)
+	return
 }