NOTE: sparsehistogram branch,

The XOR chunk refactoring made it ~2% slower to read. This PR takes it back.

This is a small step and does not fix the sloweness in some queries that we see with histogram code.

A comment with benchmark results will follow.