@codesome 2022-10-18
<details>

<summary>PromQL benchmark comparing `sparsehistogram` branch and this PR</summary>

```
benchmark                                                                                                              old ns/op      new ns/op      delta
BenchmarkRangeQuery/expr=a_one,steps=1-10                                                                              7237           7202           -0.48%
BenchmarkRangeQuery/expr=a_one,steps=10-10                                                                             7373           7305           -0.92%
BenchmarkRangeQuery/expr=a_one,steps=100-10                                                                            10694          10441          -2.37%
BenchmarkRangeQuery/expr=a_one,steps=1000-10                                                                           34593          33820          -2.23%
BenchmarkRangeQuery/expr=a_ten,steps=1-10                                                                              25866          25562          -1.18%
BenchmarkRangeQuery/expr=a_ten,steps=10-10                                                                             27336          27033          -1.11%
BenchmarkRangeQuery/expr=a_ten,steps=100-10                                                                            59032          57743          -2.18%
BenchmarkRangeQuery/expr=a_ten,steps=1000-10                                                                           291097         283338         -2.67%
BenchmarkRangeQuery/expr=a_hundred,steps=1-10                                                                          217433         216641         -0.36%
BenchmarkRangeQuery/expr=a_hundred,steps=10-10                                                                         231535         231357         -0.08%
BenchmarkRangeQuery/expr=a_hundred,steps=100-10                                                                        544371         536178         -1.51%
BenchmarkRangeQuery/expr=a_hundred,steps=1000-10                                                                       2845056        2785863        -2.08%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=1-10                                                                    10509          10695          +1.77%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=10-10                                                                   11055          11261          +1.86%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=100-10                                                                  18555          18617          +0.33%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=1000-10                                                                 79933          78695          -1.55%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1-10                                                                    31742          31542          -0.63%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=10-10                                                                   37211          37029          -0.49%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=100-10                                                                  109956         108685         -1.16%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1000-10                                                                 717937         704484         -1.87%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=1-10                                                                243631         240863         -1.14%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=10-10                                                               299226         295646         -1.20%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=100-10                                                              1023835        1010792        -1.27%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=1000-10                                                             7098318        6965222        -1.88%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=10000-10                                                                775457         767818         -0.99%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=10000-10                                                                7632520        7546190        -1.13%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=10000-10                                                            77241990       76189611       -1.36%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1-10                                                  608791         605764         -0.50%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=10-10                                                 1134913        1132111        -0.25%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=100-10                                                6328846        6327124        -0.03%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1000-10                                               58257638       58420441       +0.28%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1-10                                                  3942063        3857814        -2.14%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=10-10                                                 9142264        9087101        -0.60%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=100-10                                                61174735       61118890       -0.09%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1000-10                                               579457520      579645104      +0.03%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1-10                                              37308931       36634057       -1.81%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=10-10                                             89598897       88603933       -1.11%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=100-10                                            609073396      606936458      -0.35%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1000-10                                           5790329500     5789424042     -0.02%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=1-10                                                                 406291         398596         -1.89%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=10-10                                                                510929         505109         -1.14%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=100-10                                                               1353584        1349448        -0.31%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=1000-10                                                              9713312        9732523        +0.20%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=1-10                                                                 2816086        2750564        -2.33%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=10-10                                                                3685637        3628993        -1.54%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=100-10                                                               12095875       12032121       -0.53%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=1000-10                                                              95916705       96094580       +0.19%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1-10                                                             27524594       26865310       -2.40%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=10-10                                                            35984198       35368996       -1.71%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=100-10                                                           120101060      119562662      -0.45%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1000-10                                                          957761333      959533312      +0.19%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1-10                                                                    406146         399824         -1.56%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=10-10                                                                   511347         507271         -0.80%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=100-10                                                                  1358708        1358737        +0.00%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1000-10                                                                 9735083        9773082        +0.39%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=1-10                                                                    2820728        2761838        -2.09%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=10-10                                                                   3696894        3627678        -1.87%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=100-10                                                                  12113436       12044802       -0.57%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=1000-10                                                                 96081934       95981549       -0.10%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1-10                                                                27504942       26906185       -2.18%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=10-10                                                               36012819       35439788       -1.59%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=100-10                                                              120452972      120048694      -0.34%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1000-10                                                             960086729      963674312      +0.37%
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=1-10                                                        396855         388297         -2.16%
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=10-10                                                       457598         451692         -1.29%
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=100-10                                                      830719         833509         +0.34%
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=1000-10                                                     4416121        4447146        +0.70%
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=1-10                                                        2707406        2647520        -2.21%
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=10-10                                                       3118580        3043000        -2.42%
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=100-10                                                      6730933        6640086        -1.35%
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=1000-10                                                     42897115       42711438       -0.43%
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1-10                                                    26550959       25827721       -2.72%
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=10-10                                                   30183557       29605735       -1.91%
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=100-10                                                  66390359       65873992       -0.78%
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1000-10                                                 427134708      428694028      +0.37%
BenchmarkRangeQuery/expr=-a_one,steps=1-10                                                                             8394           8350           -0.52%
BenchmarkRangeQuery/expr=-a_one,steps=10-10                                                                            8558           8543           -0.18%
BenchmarkRangeQuery/expr=-a_one,steps=100-10                                                                           11851          11721          -1.10%
BenchmarkRangeQuery/expr=-a_one,steps=1000-10                                                                          36684          35806          -2.39%
BenchmarkRangeQuery/expr=-a_ten,steps=1-10                                                                             29281          28797          -1.65%
BenchmarkRangeQuery/expr=-a_ten,steps=10-10                                                                            30794          30305          -1.59%
BenchmarkRangeQuery/expr=-a_ten,steps=100-10                                                                           63254          61375          -2.97%
BenchmarkRangeQuery/expr=-a_ten,steps=1000-10                                                                          300459         292783         -2.55%
BenchmarkRangeQuery/expr=-a_hundred,steps=1-10                                                                         238093         234083         -1.68%
BenchmarkRangeQuery/expr=-a_hundred,steps=10-10                                                                        253936         248891         -1.99%
BenchmarkRangeQuery/expr=-a_hundred,steps=100-10                                                                       573109         559677         -2.34%
BenchmarkRangeQuery/expr=-a_hundred,steps=1000-10                                                                      2933958        2863476        -2.40%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1-10                                                                      14738          14587          -1.02%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10-10                                                                     17347          17331          -0.09%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=100-10                                                                    45248          44775          -1.05%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1000-10                                                                   300916         299207         -0.57%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=1-10                                                                      64225          63088          -1.77%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=10-10                                                                     100741         98936          -1.79%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=100-10                                                                    487293         487001         -0.06%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=1000-10                                                                   4180270        4168042        -0.29%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=1-10                                                              555128         548926         -1.12%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10-10                                                             941741         934207         -0.80%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=100-10                                                            5140937        5125051        -0.31%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=1000-10                                                           45821236       45781868       -0.09%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10000-10                                                                  2984937        2971526        -0.45%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=10000-10                                                                  42824314       42596340       -0.53%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10000-10                                                          466757403      469783805      +0.65%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1-10                                                     32483          32617          +0.41%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=10-10                                                    33118          33485          +1.11%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=100-10                                                   42671          42477          -0.45%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1000-10                                                  125342         124277         -0.85%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1-10                                                     67554          66676          -1.30%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=10-10                                                    81064          80374          -0.85%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=100-10                                                   238568         237016         -0.65%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1000-10                                                  1682732        1668699        -0.83%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1-10                                             402950         396629         -1.57%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=10-10                                            564706         557449         -1.29%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=100-10                                           2422177        2399258        -0.95%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1000-10                                          19798695       19687149       -0.56%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1-10                                                      32615          32674          +0.18%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=10-10                                                     33825          34076          +0.74%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=100-10                                                    48942          48728          -0.44%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1000-10                                                   187004         185949         -0.56%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=1-10                                                      70430          70073          -0.51%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=10-10                                                     94582          93823          -0.80%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=100-10                                                    354454         352623         -0.52%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=1000-10                                                   2825862        2818772        -0.25%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=1-10                                              433895         426408         -1.73%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=10-10                                             682141         676739         -0.79%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=100-10                                            3426581        3408050        -0.54%
…